feat(plugin-text): consistent selection after adding a new row
